requirements.txt
# IAO Architecture Pipeline — Python Dependencies
openpyxl>=3.1.5
PyYAML>=6.0
Jinja2>=3.1
requests>=2.31
python-dotenv>=1.0
Markdown>=3.6
pymdown-extensions>=10.0
fpdf2>=2.8
pandas>=2.0
